[frontend] Types in filter are now translated like in the catalog card chip
